# update to use dartdevc as a persistent worker