--- a/benchmark/bench.dart
+++ b/benchmark/bench.dart
@@ -8,32 +8,51 @@ library services.bench;
 
 import 'dart:async';
 
+import 'package:logging/logging.dart';
+
 import 'package:dart_services/src/analysis_server.dart';
 import 'package:dart_services/src/bench.dart';
 import 'package:dart_services/src/common.dart';
 import 'package:dart_services/src/compiler.dart';
 import 'package:dart_services/src/flutter_web.dart';
 
-void main(List<String> args) {
+void main(List<String> args) async {
   final bool json = args.contains('--json');
 
   final BenchmarkHarness harness = BenchmarkHarness(asJson: json);
 
   final FlutterWebManager flutterWebManager = FlutterWebManager(sdkPath);
+  await flutterWebManager.warmup();
+  var compiler = Compiler(sdkPath, flutterWebManager);
+
+  Logger.root.level = Level.WARNING;
+  Logger.root.onRecord.listen((LogRecord record) {
+    print(record);
+    if (record.stackTrace != null) print(record.stackTrace);
+  });
 
   final List<Benchmark> benchmarks = [
     AnalyzerBenchmark('hello', sampleCode, flutterWebManager),
     AnalyzerBenchmark('hellohtml', sampleCodeWeb, flutterWebManager),
     AnalyzerBenchmark('sunflower', _sunflower, flutterWebManager),
+    AnalyzerBenchmark('spinning_square', _spinningSquare, flutterWebManager),
     AnalysisServerBenchmark('hello', sampleCode, flutterWebManager),
     AnalysisServerBenchmark('hellohtml', sampleCodeWeb, flutterWebManager),
     AnalysisServerBenchmark('sunflower', _sunflower, flutterWebManager),
-    Dart2jsBenchmark('hello', sampleCode, flutterWebManager),
-    Dart2jsBenchmark('hellohtml', sampleCodeWeb, flutterWebManager),
-    Dart2jsBenchmark('sunflower', _sunflower, flutterWebManager),
+    AnalysisServerBenchmark(
+        'spinning_square', _spinningSquare, flutterWebManager),
+    Dart2jsBenchmark('hello', sampleCode, compiler),
+    Dart2jsBenchmark('hellohtml', sampleCodeWeb, compiler),
+    Dart2jsBenchmark('sunflower', _sunflower, compiler),
+    Dart2jsBenchmark('spinning_square', _spinningSquare, compiler),
+    DevCompilerBenchmark('hello', sampleCode, compiler),
+    DevCompilerBenchmark('hellohtml', sampleCodeWeb, compiler),
+    DevCompilerBenchmark('sunflower', _sunflower, compiler),
+    DevCompilerBenchmark('spinning_square', _spinningSquare, compiler),
   ];
 
-  harness.benchmark(benchmarks);
+  await harness.benchmark(benchmarks);
+  await compiler.dispose();
 }
 
 class AnalyzerBenchmark extends Benchmark {
@@ -60,14 +79,27 @@ class Dart2jsBenchmark extends Benchmark {
   final String source;
   final Compiler compiler;
 
-  Dart2jsBenchmark(
-      String name, this.source, FlutterWebManager flutterWebManager)
-      : compiler = Compiler(sdkPath, flutterWebManager),
-        super('dart2js.$name');
+  Dart2jsBenchmark(String name, this.source, this.compiler)
+      : super('dart2js.$name');
 
   @override
   Future perform() {
     return compiler.compile(source).then((CompilationResults result) {
+      if (!result.success) throw result;
+    });
+  }
+}
+
+class DevCompilerBenchmark extends Benchmark {
+  final String source;
+  final Compiler compiler;
+
+  DevCompilerBenchmark(String name, this.source, this.compiler)
+      : super('dartdevc.$name');
+
+  @override
+  Future perform() {
+    return compiler.compileDDC(source).then((DDCCompilationResults result) {
       if (!result.success) throw result;
     });
   }
@@ -106,7 +138,7 @@ class Sunflower {
   static const String ORANGE = "orange";
   static const SEED_RADIUS = 2;
   static const SCALE_FACTOR = 4;
-  static const TAU = math.PI * 2;
+  static const TAU = math.pi * 2;
   static const MAX_D = 300;
 
   CanvasRenderingContext2D ctx;
@@ -155,5 +187,57 @@ class Sunflower {
     ctx.closePath();
     ctx.stroke();
   }
+}
+''';
+
+final _spinningSquare = '''
+// Copyright 2018 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+import 'package:flutter_web/material.dart';
+import 'package:flutter_web_ui/ui.dart' as ui;
+
+class SpinningSquare extends StatefulWidget {
+  @override
+  _SpinningSquareState createState() => new _SpinningSquareState();
+}
+
+class _SpinningSquareState extends State<SpinningSquare>
+    with SingleTickerProviderStateMixin {
+  AnimationController _animation;
+
+  @override
+  void initState() {
+    super.initState();
+    // We use 3600 milliseconds instead of 1800 milliseconds because 0.0 -> 1.0
+    // represents an entire turn of the square whereas in the other examples
+    // we used 0.0 -> math.pi, which is only half a turn.
+    _animation = new AnimationController(
+      duration: const Duration(milliseconds: 3600),
+      vsync: this,
+    )..repeat();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return new RotationTransition(
+        turns: _animation,
+        child: new Container(
+          width: 200.0,
+          height: 200.0,
+          color: const Color(0xFF00FF00),
+        ));
+  }
+
+  @override
+  void dispose() {
+    _animation.dispose();
+    super.dispose();
+  }
+}
+
+main() async {
+  await ui.webOnlyInitializePlatform();
+  runApp(new Center(child: new SpinningSquare()));
 }
 ''';