--- a/benchmark/bench.dart
+++ b/benchmark/bench.dart
@@ -22,6 +22,7 @@ void main(List<String> args) async {
   final BenchmarkHarness harness = BenchmarkHarness(asJson: json);
 
   final FlutterWebManager flutterWebManager = FlutterWebManager(sdkPath);
+  // Fetches the flutter web summary files etc from google cloud storage.
   await flutterWebManager.warmup();
   var compiler = Compiler(sdkPath, flutterWebManager);