Closes https://github.com/dart-lang/dart-pad/issues/1003

cc @devoncarew @RedBrogdon @domesticmouse not sure what combination of you would like to review this :).

This uses the bazel_worker package to run a pool of persistent dartdevc tasks and schedule jobs between them. Currently it is using the default number of workers (4), but that is easily tweakable (and I don't know what makes the most sense for dartpad - if it only handles one request at a time then just 1 would be needed although it doesn't hurt to have it higher technically - it would only spin up 1 in that case still). 

Note that I did not switch to using kernel - that would be a bit more involved and doesn't really seem necessary although we could get additional wins from it.

I did a few other cleanup things as well, feel free to push back on any of those, I will leave specific comments as to why I changed them though.

I added general dev compiler benchmarks and a spinning square benchmark for all modes and got the following results for ddc:

## Before
```
[dartdevc.hello            :  181.118ms]
[dartdevc.hellohtml        :  219.658ms]
[dartdevc.sunflower        :  241.977ms]
[dartdevc.spinning_square  :  251.619ms]
```

## After

```
[dartdevc.hello            :   19.113ms]
[dartdevc.hellohtml        :   32.594ms]
[dartdevc.sunflower        :   46.399ms]
[dartdevc.spinning_square  :   41.897ms]
```