@jakemac53 2019-04-12
> This looks great to me. The only thing I want to check is whether the existing mechanism that kept dart-services from trying to start two compiles at once targeting the same location will also work here. Is there anything changing that could cause state from one compileDDC request to bleed over into another or cause two processes kicked off by _ddcDriver to fight over a shared root directory for the compile?

This does work slightly differently, but it has the same behavior. The difference is that since the [processes](https://github.com/dart-lang/dart-services/pull/387/files#diff-6f8db5c7de455147a9ca9a2f78b305c2R32) are persistent they all run with the same current working directory (current just the cwd of the main process, but we could create a different temp dir if you would like and run them there for additional safety).

Each compile call still creates its [own temp directory](https://github.com/dart-lang/dart-services/pull/387/files#diff-6f8db5c7de455147a9ca9a2f78b305c2R128), where it writes the [main.dart](https://github.com/dart-lang/dart-services/pull/387/files#diff-6f8db5c7de455147a9ca9a2f78b305c2R139), and we pass the [full path](https://github.com/dart-lang/dart-services/pull/387/files#diff-6f8db5c7de455147a9ca9a2f78b305c2R146) to that to dartdevc, as well as making the [output path](https://github.com/dart-lang/dart-services/pull/387/files#diff-6f8db5c7de455147a9ca9a2f78b305c2R143) be under that directory as well.