@devoncarew 2019-04-14
I see this issue now when using the `compileDDC` endpoint locally:

```
VM2060:52 Uncaught TypeError: Cannot read property 'main' of undefined
    at <anonymous>:52:23
    at Object.execCb (VM1379 require.js:1696)
    at Module.check (VM1379 require.js:883)
    at Module.<anonymous> (VM1379 require.js:1139)
    at VM1379 require.js:134
    at VM1379 require.js:1189
    at each (VM1379 require.js:59)
    at Module.emit (VM1379 require.js:1188)
    at Module.check (VM1379 require.js:938)
    at Module.enable (VM1379 require.js:1176)
    at Module.init (VM1379 require.js:788)
    at callGetModule (VM1379 require.js:1203)
    at Object.completeLoad (VM1379 require.js:1590)
    at HTMLScriptElement.onScriptLoad (VM1379 require.js:1717)
```

I haven't investigated to narrow down the issue.