@domesticmouse 2019-04-14
I'm guessing we should add tests on the `compileDDC` end point right about now?