@devoncarew 2019-04-14
We have some to validate that the end point produces output, but not integration tests that the output can execute and produce expected results...