@domesticmouse 2019-04-15
I just confirmed this is the change that caused the issues seen in https://github.com/dart-lang/dart-services/pull/387#issuecomment-483061815

I'm about to file a revert for this PR.