@jakemac53 2019-04-15
I would be happy to look at it if somebody can point me at some docs for how to set up the full stack for development :).