@jakemac53 2019-04-15
Actually it looks like https://github.com/dart-lang/dart-pad/blob/master/CONTRIBUTING.md has some instructions, ill see if I can work out the issue here.