@devoncarew 2019-04-15
I'm generally following those instructions. From the dart-pad directory, I run:

```
grind serve-custom-backend
```

and from the dart-services directory I run:

```
dart bin/server_dev.dart --port 8002
```