@jakemac53 2019-04-15
Ah using `dart bin/server_dev.dart --port 8002` works (instead of `grind serve` which doesn't seem to work for me).

I noticed it still seems to be using dart2js though - how do i switch it to use the dev compiler?