@devoncarew 2019-04-15
If testing with the main dartpad UI (the desktop version), you'll need to toggle the compile time flag in lib/playground.dart - `_useDDC`.