@jakemac53 2019-04-15
Ok - still getting some errors fetching the `dart_sdk.js` file (loading the `dart_sdk` module) but I was able to at least see the compiled artifact and this is almost certainly the problem:

```
// Exports:
return {
  ________________________tmp__dartpadESMJSE__main: ________________________tmp__dartpadESMJSE__main
};
```

Dartpad is just expecting that to look like this:

// Exports:
return {
  main: main
};
```

I think I know how to fix that (pass --module-root).